+++ build.py
@@ -3,7 +3,6 @@ import sys
 import os
 import shutil
 
-import util
 import pathlib
 import shutil
