+++ csv/Camera.csv
@@ -682,12 +682,12 @@ getMapBaseMtx__15CameraTargetObjCFv,CameraManPause.o,Camera.a,true
 clampBoth__2MRFPfff,CameraManPause.o,Camera.a,false
 isZeroFrameMoveOff__14CameraManPauseCFv,CameraManPause.o,Camera.a,true
 isCollisionOff__14CameraManPauseCFv,CameraManPause.o,Camera.a,true
-__ct__19CameraManSubjectiveFPCc,CameraManSubjective.o,Camera.a,false
-init__19CameraManSubjectiveFRC12JMapInfoIter,CameraManSubjective.o,Camera.a,false
-calc__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,false
-notifyActivate__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,false
-notifyDeactivate__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,false
-__dt__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,false
+__ct__19CameraManSubjectiveFPCc,CameraManSubjective.o,Camera.a,true
+init__19CameraManSubjectiveFRC12JMapInfoIter,CameraManSubjective.o,Camera.a,true
+calc__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,true
+notifyActivate__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,true
+notifyDeactivate__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,true
+__dt__19CameraManSubjectiveFv,CameraManSubjective.o,Camera.a,true
 __dt__18CameraMedianPlanetFv,CameraMedianPlanet.o,Camera.a,false
 __ct__18CameraMedianPlanetFPCc,CameraMedianPlanet.o,Camera.a,true
 reset__18CameraMedianPlanetFv,CameraMedianPlanet.o,Camera.a,false
