+++ include/Game/Camera/CameraManSubjective.h
@@ -7,6 +7,13 @@ class CameraSubjective;
 class CameraManSubjective : public CameraMan {
 public:
     CameraManSubjective(const char *);
+    virtual ~CameraManSubjective();
+    
+    virtual void init(const JMapInfoIter &);
+
+    virtual void calc();
+    virtual void notifyActivate();
+    virtual void notifyDeactivate();
 
     CameraSubjective *mCamera;  // _48
 };
